+++ src/main/kotlin/it/czerwinski/intellij/wavefront/lang/MtlSyntaxHighlighter.kt
@@ -53,6 +53,9 @@ class MtlSyntaxHighlighter : SyntaxHighlighterBase() {
             MtlTypes.TRANSPARENCY_KEYWORD,
             MtlTypes.ROUGHNESS_KEYWORD,
             MtlTypes.METALNESS_KEYWORD,
+            MtlTypes.SHEEN_KEYWORD,
+            MtlTypes.CLEARCOAT_THICKNESS_KEYWORD,
+            MtlTypes.CLEARCOAT_ROUGHNESS_KEYWORD,
             MtlTypes.AMBIENT_COLOR_MAP_KEYWORD,
             MtlTypes.DIFFUSE_COLOR_MAP_KEYWORD,
             MtlTypes.SPECULAR_COLOR_MAP_KEYWORD,
@@ -63,6 +66,7 @@ class MtlSyntaxHighlighter : SyntaxHighlighterBase() {
             MtlTypes.STENCIL_DECAL_MAP_KEYWORD,
             MtlTypes.ROUGHNESS_MAP_KEYWORD,
             MtlTypes.METALNESS_MAP_KEYWORD,
+            MtlTypes.SHEEN_MAP_KEYWORD,
             MtlTypes.NORMAL_MAP_KEYWORD,
             MtlTypes.BUMP_MAP_KEYWORD,
             MtlTypes.REFLECTION_MAP_KEYWORD ->